+++ src/utils/useAddToMapInstance.ts
@@ -1,23 +1,31 @@
+import { Map } from 'leaflet'
 import { useEffect, useState } from 'react'
 import useMapInstance from './useMapInstance'
 import { AllLeafletInstances } from '../types'
 
 export default <T extends AllLeafletInstances>(
-  instance: Partial<AllLeafletInstances>,
+  instance: T,
   setInstance: Function,
+  customAdd?: (instance: T, mapInstance: Map) => T,
 ) => {
   const mapInstance = useMapInstance()
-  const [myInstance, setMyInstance] = useState<T>()
+  const [componentInstance, setComponentInstance] = useState<
+    AllLeafletInstances
+  >()
 
   useEffect(() => {
-    if (mapInstance && !myInstance && instance.addTo) {
-      setMyInstance(instance.addTo(mapInstance) as T)
+    if (mapInstance && !componentInstance) {
+      setComponentInstance(
+        customAdd
+          ? customAdd(instance, mapInstance)
+          : instance.addTo(mapInstance),
+      )
     }
   }, [instance, mapInstance, setInstance])
 
   useEffect(() => {
-    if (setInstance && myInstance) {
-      setInstance(myInstance)
+    if (setInstance && componentInstance) {
+      setInstance(componentInstance)
     }
-  }, [myInstance, setInstance])
+  }, [componentInstance, setInstance])
 }